// api/models/fuel-properties.model.ts
export interface FuelPropertiesInputVar {
// Fuel Properties from Gasification Power Generation
    model: string;
    GasolineDensity: number;
    GasolineHigherHeatingValueMjPerKg: number;
    GasolineLowerHeatingValueMjPerKg: number;
    LightDieselDensity: number;
    LightDieselHigherHeatingValueMjPerKg: number;
    LightDieselLowerHeatingValueMjPerKg: number;
    HeavyDieselDensity: number;
    HeavyDieselHigherHeatingValueMjPerKg: number;
    HeavyDieselLowerHeatingValueMjPerKg: number;
    NaturalGasDensity: number;
    NaturalGasHigherHeatingValueMjPerKg: number;
    NaturalGasLowerHeatingValueMjPerKg: number;
    COHigherHeatingValueMjPerKg: number;
    COLowerHeatingValueMjPerKg: number;
    H2HigherHeatingValueMjPerKg: number;
    H2LowerHeatingValueMjPerKg: number;
    CH4HigherHeatingValueMjPerKg: number;
    CH4LowerHeatingValueMjPerKg: number;
  }